+++ packages/Photos/src/site/components/com_photos/views/photos/html/masonry.php
@@ -1,11 +1,22 @@
 <?php defined('KOOWA') or die('Restricted access');?>
 
 <?= @helper('ui.header') ?>
+
+<?
+$url['layout'] = 'masonry_list';
+$tags = array('hashtag', 'location', 'mention');
+foreach ($tags as $tag) {
+    if (isset(${$tag})) {
+        $url[$tag] = ${$tag};
+    }
+}
+?>
+
 <?= @infinitescroll($photos, array(
   'id' => 'an-photos',
-  'url' => 'layout=masonry_list',
+  'url' => $url,
   'layout_list' => 'masonry',
   'layout_item' => 'masonry',
   'hiddenlink' => true,
-  'columns' => 3
+  'columns' => 2
 )) ?>