+++ src/site/components/com_people/views/people/html/default.php
@@ -2,7 +2,18 @@
 
 <?= @helper('ui.header') ?>
 
+<?
+$url['layout'] = 'list';
+$tags = array('hashtag', 'location', 'mention');
+foreach ($tags as $tag) {
+    if (isset(${$tag})) {
+        $url[$tag] = ${$tag};
+    }
+}
+?>
+
 <?= @infinitescroll($people, array(
   'id' => 'an-people',
+  'url' => $url,
   'hiddenlink' => true,
 )) ?>